fix: minor MFARegistrationStateSnapshot refactor
