+++ tests/unit/components/MultifactorAuthentication/observability/trackMFAFlowOutcome.test.ts
@@ -1,5 +1,5 @@
+import type {MFARegistrationStateSnapshot} from '@components/MultifactorAuthentication/biometrics/captureRegistrationState';
 import trackMFAFlowOutcome from '@components/MultifactorAuthentication/observability/trackMFAFlowOutcome';
-import type {MFARegistrationStateSnapshot} from '@components/MultifactorAuthentication/observability/trackMFAFlowOutcome';
 
 import {createLocalMFAError} from '@libs/MultifactorAuthentication/shared/MFAResult';
